falta tratar erros da AdicionarEquipamentos

+++ app/src/main/java/com/solares/calculadorasolar/classes/Equipamentos.java
@@ -2,10 +2,10 @@ package com.solares.calculadorasolar.classes;
 
 public class Equipamentos {
     String nome;
-    int quantidade;
+    double quantidade;
     double potencia;
-    int horasPorDia;
-    int diasUtilizados;
+    double horasPorDia;
+    double diasUtilizados;
     boolean CC;
 
     public Equipamentos(){
@@ -27,16 +27,16 @@ public class Equipamentos {
     }
 
     public void setNome(String nome){ this.nome=nome;}
-    public void setQuantidade(int quantidade){ this.quantidade=quantidade;}
+    public void setQuantidade(double quantidade){ this.quantidade=quantidade;}
     public void setPotencia(double potencia){ this.potencia=potencia;}
-    public void setHorasPorDia(int horasPorDia){ this.horasPorDia=horasPorDia;}
+    public void setHorasPorDia(double horasPorDia){ this.horasPorDia=horasPorDia;}
     public void setCC(boolean CC){ this.CC = CC;}
-    public void setDiasUtilizados(int diasUtilizados){ this.diasUtilizados = diasUtilizados;}
+    public void setDiasUtilizados(double diasUtilizados){ this.diasUtilizados = diasUtilizados;}
 
     public String getNome(){ return this.nome;}
-    public int getQuantidade(){ return this.quantidade;}
+    public double getQuantidade(){ return this.quantidade;}
     public double getPotencia(){ return this.potencia;}
-    public int getHorasPorDia(){ return this.horasPorDia;}
+    public double getHorasPorDia(){ return this.horasPorDia;}
     public boolean getCC(){ return this.CC;}
-    public int getDiasUtilizados(){ return this.diasUtilizados;}
+    public double getDiasUtilizados(){ return this.diasUtilizados;}
 }
